requirements.txt
fastapi[all]==0.110.0
uvicorn==0.27.1
celery==5.3.6
redis==5.0.3
qdrant-client==1.8.0
langchain>=0.2.14,<0.3.0
langchain-community>=0.2.12,<0.3.0
langchain-ollama>=0.1.1,<0.2.0
langchain-text-splitters>=0.2.0,<0.3.0
pypdf>=4.2.0
google-generativeai==0.4.1
sentence-transformers==2.5.1
pydantic==2.6.4
